Generate the field list to use number in row.get()
