+++ examples/todo.rs
@@ -121,8 +121,6 @@ fn main() {
     // Create the table.
     let _ = sql!(TodoItem.create());
 
-    tql::init(&connection);
-
     let mut args = env::args();
     args.next();
