+++ tests/update_expr.rs
@@ -75,8 +75,6 @@ fn test_update() {
     let _ = sql!(RelatedTable.create());
     let _ = sql!(TableUpdateExpr.create());
 
-    tql::init(&connection);
-
     let id = sql!(RelatedTable.insert(field1 = "")).unwrap();
     let related_field = sql!(RelatedTable.get(id)).unwrap();
 
@@ -136,8 +134,6 @@ fn test_update_operation() {
 
     let _ = sql!(RelatedTable.create());
     let _ = sql!(TableUpdateExpr.create());
-
-    tql::init(&connection);
 
     let id = sql!(RelatedTable.insert(field1 = "")).unwrap();
     let related_field = sql!(RelatedTable.get(id)).unwrap();
